USBPort works better
It is not possible to read only a few bytes from USBPort.  The read function will always read until it fills the buffer or reaches the last byte. It is not possible to read only 1 byte for instance if there are more than that in the pipe.  So now, optotune works...!