Started working on USB4000
This will require changing the design a bit: the endpoints are quite different (there is no "main" and "secondary", they change al the time). Also, the Status is device-specific: some parameters take the place of others.  I think the best will be to create a Status class that is defined within the Spectrometer class (or use a StatusUSB2000 and StatusUSB4000, etc....)