Working USB4000
The USB4000 has several differences with the USB2000
° Status is different and incompatible
* Integration time in microseconds
* Different protocol to request a spectrum
* Different protocol to get the spectrum data

This is the first time eveyrthing works, including setting the integration time.  The strategy was to simply override the OISpectrometer function when needed, and replace isSpectrumRequested() with True (there is no such function for USB4000). 

There are some hiccups at the end of the spectrum, I think this may be incomplete data transferred to endpoints, which I wil investigate right now.