Temp commit to show some progress
acquisitionStatus is not defined in the documentation. I suspect that the first bit 0 is the same as in the USB2000, but not sure.  bit 1,2 and 3 are also sometimes set but I don't know what they are.
The number of packets transferred never goes above 6.
I created a flushEndpoints() function that cycles through all input endpoints to read their exisiting buffers and drop them.