Leading and trailing samples can be discarded, better buffers
It appears that there are some samples at the beginning and end that can be discarded (for instance, trailing samples are all the same).
We flush them in wavelength and spectrum, so they always match.
For now, it is hardcoded.  Probably a bad idea.

Buffers are now just the right size for USB.read() for each endpoint using wMaxPacketSize