That's it: I've had it with libs on Windows
To manage the mess on Windows with libraries, where they are which one to pick, I am now bundling libusb with PyHardwareLibrary.  What a fucking nightmare to deal with: can anyone tell me where the hell the standard library directories are??? Are there any? Mine is /usr/local/lib.  Why can't you have that, Windows???

So now, we validate the PyUSB backend manually to make sure it works.
A lot of people don't know what to do when installing PyHardwareLibrary and PyUSB: it is not clear how to get PyUSB to recognize the installed version.  In fact, the web is filled with crap and ways and driver replacement advice and visual studio does this and bla and whatever. What a bloody mess: nobody knows where to put them.  KnownDLLs? Really?