Create stellarnet.zip
The Stellarnet software needs a licence.  I know this is just a driver.  I have argued with Stellarnet.  So you need to communicate with them, sign a licence agreement then you can get the file from them. It is included at this point (encrypted) so I can distribute it easily to my collaborators.