We cannot derive from an Enum subclass
So I cannot create a Name subclass of Enum and use it as an argument for name by having users create a subclass of Name (like in Swift).
https://stackoverflow.com/questions/33679930/how-to-extend-python-enum

So if I simply request people to use Enum as their base class, it's ok and using the Enum as the key in the dictionary will work (contrary to when it was a subclass).  This code is therefore fixed.