Had to hack the Sutter into PyFTDI
For some reason, it is difficult to get the PyFTDI urls because the wildcard does not seem to work.  I found a way to get the ftdi:// for the sutter, which I added to SerialPort.  It is hardcoded, but only in SerialPort.  SutterDevice() is not affected and now uses a general approach.