Many modificaions to PhysicalDevice and subclasses to manage instantiation
Many aspects of instantiations are general: idProduct, idVendor must match and can be validated before initialize().  The communication channel is only opened in doInitializeDevice(): a PhysicalDevice can be created even if it is not connected, the initializeDevice will simply fail.

To allow DeviceManager() to instantiate any class, they must have a __init__ that is standard.  Hence, all __init__ have been modified to include these three named parameters.

Also, for historical reasons, productId, and vendorId have been changed to idProduct and idVendor.

Finally, the DeviceManager(), upon recognizing that a device was connected, will attempt to match a subclass of PhysicalDevice: this is a two-step process: 1) __init__ (which can raise an incompatible class exception and 2) initialize that will attempt to communicate with the device.  Devices cannot be initialized automatically because they will lock the communication channel for all  instances, preventing any communications even for USB diagnostic purposes.